ml_pipeline/requirements.txt
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
transformers==4.37.0
sentence-transformers==2.3.1
torch==2.2.0
faiss-cpu==1.7.4
scikit-surprise==1.1.3
kagglehub==0.2.7
requests==2.31.0
pillow==10.2.0
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
shap==0.44.0
networkx==3.2.1
prophet==1.1.5
umap-learn==0.5.5
python-dotenv==1.0.0
tqdm==4.66.1
onnx==1.15.0
huggingface-hub==0.20.2
